Fix buffer overflow due to integer promotion rules
Integer promotion rules convert both sides of the comparison to float,
meaning the loss of precision on the float is not caught. This uncaught
imprecision can then result in width vs. size.width and height vs.
size.height being off by one for e.g. 2^24 + 1.

+++ src/OFQOIImageFormatHandler.m
@@ -59,7 +59,7 @@ hashPixel(uint8_t pixel[4])
 	size.width = width;
 	size.height = height;
 
-	if (size.width != width || size.height != height)
+	if ((uint32_t)size.width != width || (uint32_t)size.height != height)
 		@throw [OFOutOfRangeException exception];
 
 	switch ([stream readInt8]) {
